+++ src/spice2x/CMakeLists.txt
@@ -615,6 +615,12 @@ set(SOURCE_FILES ${SOURCE_FILES}
         overlay/windows/popn_sub.cpp
         overlay/windows/wnd_manager.cpp
 
+        # patcher
+        patcher/config.cpp
+        patcher/loader.cpp
+        patcher/patch_manager.cpp
+        patcher/runtime.cpp
+
         # external
         external/easywsclient/easywsclient.cpp
